requirements.txt
torch
transformers
accelerate
peft
numpy
PyYAML
tqdm
bitsandbytes